+++ packages/sdk/src/common/domain.ts
@@ -1,6 +1,5 @@
 import type * as ApiClient from "@rarible/api-client"
 import type { Action } from "@rarible/action"
-import type { FlowContractAddress } from "@rarible/flow-sdk/build/common/flow-address"
 
 // @todo draft. probably will be changed in future
 export type CurrencyType = {
@@ -17,20 +16,10 @@ export interface AbstractPrepareResponse<Id, In, Out> {
 export type RequestCurrency = ApiClient.CurrencyId | RequestCurrencyAssetType
 
 export type RequestCurrencyAssetType =
-  | ApiClient.EthErc20AssetType
-  | ApiClient.EthEthereumAssetType
-  | ApiClient.FlowAssetTypeNft
-  | ApiClient.FlowAssetTypeFt
-  | ApiClient.TezosXTZAssetType
-  | ApiClient.TezosFTAssetType
-  | ApiClient.SolanaNftAssetType
-  | ApiClient.SolanaSolAssetType
-
-export type AbstractItemId<T extends ApiClient.Blockchain, ContractAddress extends string> = {
-	blockchain: T
-	contract: ContractAddress
-	itemId: string
-}
-
-//todo completelgy get rid of this, do not parse ids
-export type FlowItemId = AbstractItemId<ApiClient.Blockchain.FLOW, FlowContractAddress>
+	| ApiClient.EthErc20AssetType
+	| ApiClient.EthEthereumAssetType
+	| ApiClient.FlowAssetTypeFt
+	| ApiClient.TezosXTZAssetType
+	| ApiClient.TezosFTAssetType
+	| ApiClient.SolanaNftAssetType
+	| ApiClient.SolanaSolAssetType