+++ packages/sdk/src/sdk-blockchains/solana/tests/mint.test.ts
@@ -8,15 +8,15 @@ import { retry } from "../../../common/retry"
 
 describe("Solana mint", () => {
 	const wallet = getWallet()
-	const sdk = createRaribleSdk(new SolanaWallet(wallet), "dev", { logs: LogsLevel.DISABLED })
+	const sdk = createRaribleSdk(new SolanaWallet(wallet), "development", { logs: LogsLevel.DISABLED })
 
 	test("mint an nft", async () => {
 		const { submit } = await sdk.nft.mint({
-			collectionId: toCollectionId("SOLANA:65DNtgn5enhi6QXevn64jFq41Qgv71bvr8UVVwGiYkLJ"),
+			collectionId: toCollectionId("SOLANA:Ev9n3xAfCrxPrUSUN4mLorwfaknjj4QMcyLUnbPymSmJ"),
 		})
 
 		const res = await submit({
-			supply: 1,
+			supply: 0,
 			lazyMint: false,
 			uri: "https://arweave.net/Vt0uj2ql0ck-U5dLWDWJnwQaZPrvqkfxils8agrTiOc",
 		})