+++ packages/sdk/src/sdk-blockchains/tezos/test/test-wallet.ts
@@ -8,8 +8,8 @@ export function createTestWallet(edsk: string) {
 			edsk,
 			// "https://hangzhou.tz.functori.com"
 			// "https://hangzhounet.smartpy.io"
-			// "https://test-tezos-node.rarible.org"
 			"https://test-tezos-node.rarible.org"
+			// "https://dev-tezos-node.rarible.org"
 		)
 	)
 }